Fixed overflow in import IO throughput report

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/UpdateRecordsStep.java
@@ -42,7 +42,7 @@ public class UpdateRecordsStep<RECORD extends AbstractBaseRecord>
 {
     protected final RecordStore<RECORD> store;
     private final int recordSize;
-    private int recordsUpdated;
+    private long recordsUpdated;
 
     public UpdateRecordsStep( StageControl control, Configuration config, RecordStore<RECORD> store )
     {
